Revert "Modularize monolithic app.py into maintainable architecture (2816 → 169 lines)"
