--- streamlit_app/ai/__init__.py
@@ -1,27 +1,0 @@
-# AI module for machine learning models and predictions
-
-from .smart_entry import (
-    calculate_smart_entry_price, 
-    display_smart_signal_results
-)
-from .predictor import predict_with_models
-from .model_trainer import (
-    train_simple_lstm,
-    train_simple_xgboost, 
-    train_simple_cnn,
-    train_simple_svc,
-    train_simple_naive_bayes,
-    train_and_save_all_models
-)
-
-__all__ = [
-    'calculate_smart_entry_price',
-    'display_smart_signal_results', 
-    'predict_with_models',
-    'train_simple_lstm',
-    'train_simple_xgboost',
-    'train_simple_cnn', 
-    'train_simple_svc',
-    'train_simple_naive_bayes',
-    'train_and_save_all_models'
-]
