--- streamlit_app/config/__init__.py
@@ -1,28 +1,0 @@
-# Configuration module for Gold Bot Trading AI
-
-from .settings import *
-
-__all__ = [
-    'SYMBOL_OPTIONS',
-    'INTERVAL_MAPPING', 
-    'MODEL_DIRECTORY',
-    'XGB_FEATURE_NAMES',
-    'CNN_FEATURE_NAMES',
-    'SVC_FEATURE_NAMES',
-    'NB_FEATURE_NAMES',
-    'DEFAULT_BALANCE',
-    'DEFAULT_RISK_PERCENT',
-    'DEFAULT_SL_PIPS',
-    'DEFAULT_TP_PIPS',
-    'LABEL_MAP',
-    'ERROR_MESSAGES',
-    'PAGE_CONFIG',
-    'APP_TITLE',
-    'APP_VERSION',
-    'NEW_SYSTEM_AVAILABLE',
-    'UTILS_AVAILABLE',
-    'WEBSOCKET_AVAILABLE',
-    'BINANCE_AVAILABLE',
-    'VALIDATION_UTILS_AVAILABLE',
-    'TF_AVAILABLE'
-]
