--- streamlit_app/data/__init__.py
@@ -1,8 +1,0 @@
-# Data module for data loading and processing
-
-from .data_loader import (
-    get_binance_data,
-    load_and_process_data_enhanced
-)
-
-__all__ = ['get_binance_data', 'load_and_process_data_enhanced']
