--- streamlit_app/ui/__init__.py
@@ -1,18 +1,0 @@
-# UI module for Streamlit interface components
-
-from .components import (
-    display_sidebar_controls,
-    display_system_status,
-    display_market_status
-)
-from .trading_interface import display_trading_interface
-from .training_interface import display_training_interface, display_model_management
-
-__all__ = [
-    'display_sidebar_controls',
-    'display_system_status', 
-    'display_market_status',
-    'display_trading_interface',
-    'display_training_interface',
-    'display_model_management'
-]
