--- streamlit_app/utils/__init__.py
@@ -1,16 +1,0 @@
-# Utility functions and formatters
-
-from .formatters import sanitize_filename, format_price
-from .validators import validate_trading_inputs
-from .model_loader import load_all_models, check_model_availability, get_model_info
-from .conversion import get_conversion_rate
-
-__all__ = [
-    'sanitize_filename', 
-    'format_price', 
-    'validate_trading_inputs',
-    'load_all_models',
-    'check_model_availability',
-    'get_model_info',
-    'get_conversion_rate'
-]
