--- streamlit_app/websockets/__init__.py
@@ -1,6 +1,0 @@
-# WebSocket module for real-time trading data
-
-from .twelve_data_client import TwelveDataWebSocketClient
-from .websocket_manager import EnhancedWebSocketManager
-
-__all__ = ['TwelveDataWebSocketClient', 'EnhancedWebSocketManager']
